+++ CHANGELOG.md
@@ -9,10 +9,13 @@ Polyverse Boost ReST API (Backend)
 - N/A
 
 ### Enhancements
-- N/A
+- Enable grooming/deletion of Assistants and Files to filter by username or org if the user is not an admin
+- Enable grooming/deletion of Assistants and Files to delete across users if the user is an admin
 
 ### Bug Fixes
-- N/A
+- Fix OpenAI throttling issue with Assistant batch delete (wasn't waiting 1 second)
+- Fix search/filter checks for Assistant and File Searches/Deletions
+    - Was not correctly filtering based on email, org, project - and deleted too many files
 
 ## Version 0.12.2: February 29th, 2024